+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] - 2026-03-08
+
+### Added
+- Add a `Citation Changes` tab to show newly detected citation relationships between PEPs.
+- Add data documentation ( `data/README.md` ) and a `Data` link in the dashboard header.
+
+### Changed
+- Detect and record changes in citation relationships during the data preprocessing step.
+- Show hover descriptions for the node size options in the `Network` tab.
+
 ## [0.4.0] - 2026-03-03
 
 ### Added
@@ -16,7 +26,7 @@ All notable changes to this project will be documented in this file.
 ## [0.3.0] - 2026-02-28
 
 ### Added
-- Added the PEP Metrics tab.
+- Added the `PEP Metrics` tab.
 - Enabled node size scaling by PageRank in the Network tab.
 
 ### Changed
@@ -30,7 +40,7 @@ All notable changes to this project will be documented in this file.
 ## [0.2.0] - 2026-02-26
 
 ### Added
-- Added the Network tab.
+- Added the `Network` tab.
 - Added links to the User Guide in the header and within each tab.
 - Added links to the Changelog in the header.
 - Added scheduled data update processing.